pyproject.toml
[tool.poetry]
name = "ledgertools"
version = "0.1.0"
description = "Ledger Tools"
authors = ["John Mee"]

[tool.poetry.scripts]
upbank = "ledgertools.upbank:cli"
accounts = "ledgertools.accounts:cli"

[tool.poetry.dependencies]
click = "latest"
python = "^3.10"
PyYAML = "latest"
requests = "latest"
beancount = "^2.3.5"

[tool.poetry.dev-dependencies]
pytest = "latest"
black = "latest"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
